backend/requirements.txt
pandas
numpy
wearipedia
psycopg2-binary
python-dotenv
requests
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
databases
python-dotenv
pydantic-settings